Slash commands rework

+++ bot.py
@@ -1,5 +1,7 @@
+import re
 import discord
 import logging
+import logging.handlers
 import json
 import os
 import sys
@@ -7,26 +9,71 @@ import time
 import atexit
 import random
 import linecache
-import logging
+import argparse
+from os.path import join, split
 
-client = discord.Client()
-with open("token.json", "r") as f:
+ROOT_DIR = split(__file__)[0]
+
+parser = argparse.ArgumentParser("Keabot", "Run to start the discord bot Keabot. Tracks 'reddit gold', allows image upload and random reposting.")
+parser.add_argument("-t", dest="token_file", required=False)
+parser.add_argument("-d", dest="dataFolder", default=join(ROOT_DIR, "./Data/"), required=False)
+parser.add_argument("-c", dest="configLoc", default=join(ROOT_DIR, "./config.json"), required=False)
+
+args = parser.parse_args()
+TOKEN_FILE = args.token_file
+DATA_DIR = os.path.abspath(args.dataFolder)
+IMAGES_DIR = join(DATA_DIR, "images")
+CONFIG_LOC = os.path.abspath(args.configLoc)
+prefix = ".."
+TOKEN = None
+
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.DEBUG)
+formatter = logging.Formatter('[%(asctime)s]:[%(levelname)s] %(message)s')
+
+streamHandler = logging.StreamHandler()
+streamHandler.setLevel(logging.DEBUG)
+streamHandler.setFormatter(formatter)
+logger.addHandler(streamHandler)
+
+fileHandler = logging.handlers.RotatingFileHandler(
+    join(DATA_DIR, "logs", "keabot.log"), maxBytes=(1048576*5), backupCount=7)
+fileHandler.setFormatter(formatter)
+logger.addHandler(fileHandler)
+
+with open(CONFIG_LOC, "r") as f:
     configs = json.load(f)
-    discordToken = configs["discord"]
+intents = discord.Intents.default()
+intents.message_content = True
+client = discord.Client(intents=intents)
+if TOKEN_FILE:
+    logger.info("Opening token file %s", TOKEN_FILE)
+    try:
+        with open(TOKEN_FILE, "r", encoding="utf-8") as fp:
+            TOKEN = fp.read()
+    except:
+        logger.critical("Failed to open file: %s", TOKEN_FILE)
 
+if not TOKEN:
+    TOKEN = os.environ.get("DISCORD_TOKEN")
 
-
+if not TOKEN:
+    logger.error("No token specified. Provide on command line or in environment")
+    sys.exit(1)
+else:
+    logger.info("Token is %s", TOKEN)
 #image function support
 lastSentFile = ""
-CURR_DIR = os.getcwd()
 
-#
 #database setup
-if os.access("clownScore.json", os.F_OK):
-    with open("clownScore.json", "r") as read_file:
+score_filename = join(ROOT_DIR, "clownScore.json")
+if os.access(score_filename, os.F_OK):
+    with open(score_filename, "r") as read_file:
         scoreDB = json.load(read_file)
-    backupName = time.strftime("backup/clownScore_%Y_%m_%d_%H_%M.json")
-    with open(backupName, "w") as f:
+    backupName = time.strftime("clownScore_%Y_%m_%d_%H_%M.json")
+    with open(
+        join(DATA_DIR, "backup", backupName),
+        "w") as f:
         json.dump(scoreDB, f)
     
 else:
@@ -35,8 +82,8 @@ else:
 
 clowns = {}
 
-files = os.listdir(CURR_DIR + "/Images/")
-for f in files:
+folders = os.listdir(IMAGES_DIR)
+for f in folders:
     clowns[f] = True
 with open("clowns.json", "w") as f:
     json.dump(clowns, f)
@@ -66,17 +113,17 @@ def PrintException():
 
 @client.event
 async def on_ready():
-    print('We have logged in as {0.user}'.format(client))
-    print(CURR_DIR)
+    logger.info('We have logged in as %s', client.user)
+    logger.info("ROOT_DIR: %s", ROOT_DIR)
 
 @client.event
-async def on_message(message):
+async def on_message(message: discord.Message):
     
     if message.author == client.user or message.author.bot:
-        print("Sender was me or bot")
+        logger.debug("Sender was me or bot")
         return
     if not message.content.startswith(prefix):
-        print("Sender did not use ..")
+        logger.debug("Sender did not use ..")
         return
     
     args = message.content.lower().strip()[len(prefix):].split(' ')
@@ -84,10 +131,12 @@ async def on_message(message):
     server = str(message.guild.id) 
     user = str(message.author.id) 
 
-
+    logger.info("Message: %s", message)
+    logger.info("Message content: %s", message.content)
     #new server handler
     if server not in scoreDB:
         scoreDB[server] = {}
+        logger.info("New server created in DB: %s", server)
     #
 
     #reddit gold commands
@@ -123,20 +172,20 @@ async def on_message(message):
     #
     
 
-@client.event
-async def on_message_delete(message):
-    await message.channel.send(":eyes:")
 
 @client.event
 async def on_reaction_add(reaction, user):
-    if reaction.custom_emoji and reaction.emoji.name != "gold":
-        print("Not gold emoji")
+    if reaction.is_custom_emoji() and reaction.emoji.name != "gold":
+        logger.debug("Not gold emoji")
         return
     receiver = str(reaction.message.author.id) 
     server = str(reaction.message.guild.id) 
     gifter = str(user.id) 
     
-    
+    logger.info("Gold reaction event. Gifter: %s, Receiver: %s, Server: %s", gifter, receiver, server)
+    if server not in scoreDB:
+        scoreDB[server] = {}
+        logger.info("Created server object in scoreDB for %s", server)
     #new user handling
     if gifter not in scoreDB[server]:
         scoreDB[server][gifter] = {}
@@ -153,8 +202,7 @@ async def on_reaction_add(reaction, user):
 
     #check for selfish
     if gifter == receiver:
-        print("%s gave themself gold" % gifter)
-        print("gave self gold")
+        logger.info("%s gave themself gold", gifter)
         scoreDB[server][gifter]["self"] += 1
         await backup()
         return 
@@ -168,32 +216,37 @@ async def on_reaction_add(reaction, user):
     scoreDB[server][gifter]["name"] = user.display_name
 
     await backup()
-    print("%s gave %s gold" % (gifter, receiver))
+    logger.info("%s gave %s gold", gifter, receiver)
 
 @client.event
 async def on_reaction_remove(reaction, user):
-    if reaction.custom_emoji and reaction.emoji.name != "gold":
-        print("Not gold emoji")
+    if reaction.is_custom_emoji() and reaction.emoji.name != "gold":
+        logger.debug("Not gold emoji")
         return
     receiver = str(reaction.message.author.id )
     server = str(reaction.message.guild.id )
     taker = str(user.id )
     
+    if server not in scoreDB:
+        scoreDB[server] = {}
+        logger.info("Add new server to scoreDB for %s", server)
     
     #new user handling
     if taker not in scoreDB[server]:
         scoreDB[server][taker]["score"] = 0
         scoreDB[server][taker]["given"] = 0
         scoreDB[server][taker]["self"] = 0
+        return
     if receiver not in scoreDB[server]:
         scoreDB[server][receiver]["score"] = 0
         scoreDB[server][receiver]["given"] = 0
         scoreDB[server][receiver]["self"] = 0
+        return
     #
 
     #check for selfish
     if taker == receiver:
-        print("%s took gold from themself" % taker)
+        logger.info("%s took gold from themself", taker)
         scoreDB[server][taker]["self"] -= 1
         await backup()
         return 
@@ -204,12 +257,11 @@ async def on_reaction_remove(reaction, user):
     scoreDB[server][taker]["given"] -= 1
     #
     await backup()
-    print("%s took gold from %s" % (taker, receiver))  
+    logger.info("%s took gold from %s", taker, receiver)  
 
 
 async def leaderboard(message):
-    print()
-    print("python leaderboard called")
+    logger.info("leaderboard called")
     server = str(message.guild.id) 
     scoreboard = discord.Embed(title="Scoreboard", color=discord.Color.from_rgb(255, 0, 0))
     i = 1
@@ -233,76 +285,99 @@ async def leaderboard(message):
 async def getScore(message):
     server = str(message.guild.id) 
     score = scoreDB[server][str(message.author.id)]["score"]
-    await message.channel.send("Your score is: " + str(score))
+    await message.reply("Your score is: " + str(score))
     return
 
-async def addimage(message, args):
+async def addimage(message: discord.Message, args):
     server = str(message.guild.id) 
     if len(args) == 0:
-        await message.channel.send("No clown given, bozo")
+        logger.info("No argument was given for addimage")
+        await message.reply("No clown given, bozo")
         return
     clownName = args[0]
-    print(clownName)
-    print(clowns)
+    logger.info("Adding image for %s", clownName)
     if clownName not in clowns:
-        print("Tried to add an image to a clown that doesn't exist")
-        await message.channel.send("That clown doesn't exist")
+        logger.warning("Tried to add an image to a clown that doesn't exist")
+        await message.reply("That clown doesn't exist")
         return
     fileNonce = time.strftime("_%Y_%m_%d_%H_%M_%S_")
+    if len(message.attachments) > 30:
+        logger.warning("User attempted to upload more than 30 (%s)", str(len(message.attachments)))
+        await message.reply("Woah! Don't upload so many at once!")
+        return
     i = 0
     for image in message.attachments:
+        if image.size > pow(10, 8): # 100MB
+            logger.warning("User attempted to upload a more than 100MB file (%s)", str(image.size/1000) + "KB")
+            await message.reply("One of these files was way too big. Stopping. (%s)" % str(image.size/1000) + "KB")
+            return
         extension = image.filename[-4:]
-        path = CURR_DIR + "/Images/" + clownName + "/" + clownName + fileNonce + str(i) + extension
+        path = join(
+            IMAGES_DIR,
+            clownName,
+            clownName + fileNonce + str(i) + extension)
         await image.save(path)
-        await message.channel.send("Image saved")
+        
         i += 1
-
+    await message.reply("%s image(s) saved." % str(i))
     return
 
-async def delete(message, args):
+async def delete(message: discord.Message, args):
     server = str(message.guild.id)
-    os.remove(lastSentFile)
-
+    if not lastSentFile:
+        logger.info("Attempt to delete with no previous post")
+        await message.reply("No previous image post found")
+        return
+    try:
+        os.remove(lastSentFile)
+    except:
+        logger.error("Delete image error.")
+        await message.reply("Failed to delete")
+        return
+    logger.info("Deleted %s", lastSentFile)
+    await message.reply("Image deleted")
     return
 
 async def postImage(message, name):
+    global lastSentFile
     server = str(message.guild.id) 
     
-    path = CURR_DIR + "/Images/" + name
+    path = join(IMAGES_DIR, name)
     if os.access(path, os.F_OK):
         files = os.listdir(path)
-        print(files)
+        if len(files) < 1:
+            logger.warning("Attempt to post for user with no images")
+            await message.reply("No images for this clown exist")
+            return
         n = random.randint(0, len(files)-1)
         fileToSend = files[n]
-        fileToSend = path + "/" + fileToSend
+        fileToSend = join(path, fileToSend)
         with open(fileToSend, "rb") as f:
             fpVar = discord.File(f)
             await message.channel.send(file=fpVar)
-        print("Sent file: " + fileToSend)
+        logger.info("Sent file: %s", fileToSend)
         lastSentFile = fileToSend
     else:
-        await message.channel.send("This clown doesn't exist")
+        await message.reply("This clown doesn't exist")
     return
 
 async def addClown(message, args):
     server = str(message.guild.id) 
-    if os.access("Images/" + args[0], os.F_OK):
-        await message.channel.send("This clown already exists")
+    if os.access(join(IMAGES_DIR, args[0]), os.F_OK):
+        await message.reply("This clown already exists")
     else:
-        os.mkdir(CURR_DIR + "/Images/" + args[0])
-        print("Created directory for clown " + args[0])
-        await message.channel.send("Clown created: " + args[0])
+        os.mkdir(join(IMAGES_DIR, args[0]))
+        logger.info("Created directory for clown %s", args[0])
+        await message.reply("Clown created: " + args[0])
         clowns[args[0]] = True
     await backup()
     return
 
 async def listClowns(message):
     server = str(message.guild.id) 
-    print("..clowns called")
+    logger.info("listClowns called")
     scoreboard = discord.Embed(title="Scoreboard", color=discord.Color.from_rgb(255, 0, 0))
-    
-    path = CURR_DIR + "/Images"
-    files = os.listdir(path)
+    files = os.listdir(IMAGES_DIR)
     i = 1
     for f in files:
         scoreboard.add_field(name = str(i)+".", value = f, inline=False)
@@ -313,9 +388,10 @@ async def listClowns(message):
 
 async def removeClown(message, args):
     server = str(message.guild.id) 
+    logger.warning("Attempt to use unfinished removeClown function")
+    await message.reply("This is not implemented yet.")
     return
 
 
 
-
-client.run(discordToken)
+client.run(TOKEN)
